Reorganize tests and cleanup code
