+++ appserver/tests/application/src/test/java/org/glassfish/main/test/app/web/mrjar/MultiReleaseServletContainerInitializerTest.java
@@ -18,20 +18,13 @@ package org.glassfish.main.test.app.web.mrjar;
 
 import jakarta.servlet.ServletContainerInitializer;
 
-import java.io.BufferedReader;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.nio.file.Files;
-import java.util.jar.Manifest;
 
 import org.glassfish.main.itest.tools.GlassFishTestEnvironment;
-import org.glassfish.main.itest.tools.asadmin.Asadmin;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.asset.Asset;
-import org.jboss.shrinkwrap.api.asset.ByteArrayAsset;
 import org.jboss.shrinkwrap.api.exporter.ZipExporter;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
@@ -40,14 +33,9 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInfo;
 import org.junit.jupiter.api.condition.EnabledForJreRange;
-import org.objectweb.asm.ClassWriter;
-import org.objectweb.asm.MethodVisitor;
-import org.objectweb.asm.Type;
 
 import static java.lang.System.Logger.Level.INFO;
 import static java.lang.System.Logger.Level.WARNING;
-import static java.util.jar.Attributes.Name.MANIFEST_VERSION;
-import static java.util.jar.Attributes.Name.MULTI_RELEASE;
 import static org.glassfish.main.itest.tools.asadmin.AsadminResultMatcher.asadminOK;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -55,17 +43,10 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.condition.JRE.JAVA_11;
 import static org.junit.jupiter.api.condition.JRE.JAVA_16;
 import static org.junit.jupiter.api.condition.JRE.JAVA_17;
-import static org.objectweb.asm.Opcodes.ACC_PUBLIC;
-import static org.objectweb.asm.Opcodes.ACC_SUPER;
-import static org.objectweb.asm.Opcodes.ALOAD;
-import static org.objectweb.asm.Opcodes.ARETURN;
-import static org.objectweb.asm.Opcodes.INVOKESPECIAL;
-import static org.objectweb.asm.Opcodes.RETURN;
 import static org.objectweb.asm.Opcodes.V11;
 import static org.objectweb.asm.Opcodes.V17;
-import static org.objectweb.asm.Opcodes.V1_8;
 
-public class MultiReleaseServletContainerInitializerTest {
+public class MultiReleaseServletContainerInitializerTest extends MultiReleaseTestBase {
 
     private static final System.Logger LOG = System.getLogger(MultiReleaseServletContainerInitializerTest.class.getName());
 
@@ -73,13 +54,9 @@ public class MultiReleaseServletContainerInitializerTest {
 
     private static final String CONTEXT_ROOT = "/" + APP_NAME;
 
-    private static final String MR_LIB_FILE_NAME = "mrlib.jar";
-
     private static final String SCE_LIB_FILE_NAME = "scelib.jar";
 
     private static final String WAR_FILE_NAME = APP_NAME + ".war";
-
-    private static final Asadmin ASADMIN = GlassFishTestEnvironment.getAsadmin();
 
     @BeforeAll
     public static void deploy() throws IOException {
@@ -109,6 +86,7 @@ public class MultiReleaseServletContainerInitializerTest {
         try {
             assertAll(
                 () -> assertThat(connection.getResponseCode(), equalTo(200)),
+                // Check version of loaded class file
                 () -> assertThat(Integer.parseInt(readResponse(connection)), equalTo(V11))
             );
         } finally {
@@ -125,6 +103,7 @@ public class MultiReleaseServletContainerInitializerTest {
         try {
             assertAll(
                 () -> assertThat(connection.getResponseCode(), equalTo(200)),
+                // Check version of loaded class file
                 () -> assertThat(Integer.parseInt(readResponse(connection)), equalTo(V17))
             );
         } finally {
@@ -132,30 +111,17 @@ public class MultiReleaseServletContainerInitializerTest {
         }
     }
 
-    private String readResponse(HttpURLConnection connection) throws Exception {
-        StringBuilder response = new StringBuilder();
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(connection.getInputStream()))) {
-            String line;
-            while ((line = reader.readLine()) != null) {
-                response.append(line);
-            }
-        }
-        return response.toString();
-    }
-
     private static File createDeployment() throws IOException {
-        JavaArchive mrLib = ShrinkWrap.create(JavaArchive.class, MR_LIB_FILE_NAME)
-            .setManifest(generateManifest())
-            .addClass(Version.class)
-            .add(generateVersionClass(V1_8), packageJarEntryFor(VersionImpl.class, V1_8), classFileNameFor(VersionImpl.class))
-            .add(generateVersionClass(V11), packageJarEntryFor(VersionImpl.class, V11), classFileNameFor(VersionImpl.class))
-            .add(generateVersionClass(V17), packageJarEntryFor(VersionImpl.class, V17), classFileNameFor(VersionImpl.class));
+        // Create multi-release JAR
+        JavaArchive mrLib = createMultiReleaseLibrary();
 
+        // Create ServletContainerInitializer provider library
         JavaArchive sceLib = ShrinkWrap.create(JavaArchive.class, SCE_LIB_FILE_NAME)
             .addClass(MultiReleaseServlet.class)
             .addClass(MultiReleaseServletContainerInitializer.class)
             .addAsServiceProvider(ServletContainerInitializer.class, MultiReleaseServletContainerInitializer.class);
 
+        // Create test web application
         WebArchive webArchive = ShrinkWrap.create(WebArchive.class)
             .addAsLibrary(mrLib)
             .addAsLibrary(sceLib);
@@ -165,66 +131,5 @@ public class MultiReleaseServletContainerInitializerTest {
         webArchive.as(ZipExporter.class).exportTo(warFile, true);
         tmpDir.deleteOnExit();
         return warFile;
-    }
-
-    private static Asset generateManifest() throws IOException {
-        Manifest manifest = new Manifest();
-        manifest.getMainAttributes().put(MANIFEST_VERSION, "1.0");
-        manifest.getMainAttributes().put(MULTI_RELEASE, String.valueOf(Boolean.TRUE));
-        try (ByteArrayOutputStream out = new ByteArrayOutputStream()) {
-            manifest.write(out);
-            return new ByteArrayAsset(out.toByteArray());
-        }
-    }
-
-    private static String classFileNameFor(Class<?> c) {
-        return c.getSimpleName() + ".class";
-    }
-
-    private static String packageJarEntryFor(Class<?> c, int version) {
-        StringBuilder jarTarget = new StringBuilder();
-        switch (version) {
-            case V11:
-                jarTarget.append("META-INF/versions/11/");
-                break;
-            case V17:
-                jarTarget.append("META-INF/versions/17/");
-                break;
-            default:
-                break;
-        }
-        jarTarget.append(c.getPackageName().replace('.', '/'));
-        return jarTarget.toString();
-    }
-
-    private static Asset generateVersionClass(int version) {
-        ClassWriter cw = new ClassWriter(0);
-
-        cw.visit(version,
-            ACC_PUBLIC + ACC_SUPER,
-            Type.getInternalName(VersionImpl.class),
-            null,
-            "java/lang/Object",
-            new String[] {Type.getInternalName(Version.class)});
-
-        MethodVisitor mv = cw.visitMethod(ACC_PUBLIC, "<init>", "()V", null, null);
-        mv.visitCode();
-        mv.visitVarInsn(ALOAD, 0);
-        mv.visitMethodInsn(INVOKESPECIAL, "java/lang/Object", "<init>", "()V", false);
-        mv.visitInsn(RETURN);
-        mv.visitMaxs(1, 1);
-        mv.visitEnd();
-
-        mv = cw.visitMethod(ACC_PUBLIC, "getVersion", "()Ljava/lang/String;", null, null);
-        mv.visitCode();
-        mv.visitVarInsn(ALOAD, 0);
-        mv.visitLdcInsn(String.valueOf(version));
-        mv.visitInsn(ARETURN);
-        mv.visitMaxs(2, 1);
-        mv.visitCode();
-
-        cw.visitEnd();
-
-        return new ByteArrayAsset(cw.toByteArray());
     }
 }
